# docs-site/docs/api/openapi/coinline.openapi.json
{
  "openapi": "3.0.1",
  "info": {
    "title": "Host Platform API",
    "version": "v1"
  },
  "paths": {
...
    "/api/firmware/packages": {
      "get": {
        "tags": [
          "Firmware"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Firmware"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/PackageDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/PackageDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/firmware/versions": {
      "get": {
        "tags": [
          "Firmware"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Firmware"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/FirmwareVersionCreateDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/FirmwareVersionCreateDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/FirmwareVersionCreateDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/firmware/packages/{id}": {
      "get": {
        "tags": [
          "Firmware"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/firmware/packages/{id}/artifacts": {
      "post": {
        "tags": [
          "Firmware"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ArtifactDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/firmware/packages/{id}/compatibility-rules": {
      "post": {
        "tags": [
          "Firmware"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CompatibilityRuleDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CompatibilityRuleDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CompatibilityRuleDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/firmware/targets": {
      "get": {
        "tags": [
          "Firmware"
        ],
        "responses": {
          "200": {
            "description": "OK",
            "content": {
              "text/plain": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              },
              "application/json": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              },
              "text/json": {
                "schema": {
                  "type": "array",
                  "items": { }
                }
              }
            }
          }
        }
      },
      "post": {
        "tags": [
          "Firmware"
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/TargetDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/TargetDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
    "/api/firmware/jobs/{id}": {
      "get": {
        "tags": [
          "Firmware"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/firmware/jobs/{id}/simulate": {
      "post": {
        "tags": [
          "Firmware"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/firmware/jobs/{id}/approve": {
      "post": {
        "tags": [
          "Firmware"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/ApproveDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
    "/api/firmware/jobs/{id}/cancel": {
      "post": {
        "tags": [
          "Firmware"
        ],
        "parameters": [
          {
            "name": "id",
            "in": "path",
            "required": true,
            "schema": {
              "type": "string",
              "format": "uuid"
            }
          }
        ],
        "requestBody": {
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelJobDto"
              }
            },
            "text/json": {
              "schema": {
                "$ref": "#/components/schemas/CancelJobDto"
              }
            },
            "application/*+json": {
              "schema": {
                "$ref": "#/components/schemas/CancelJobDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": "OK"
          }
        }
      }
    },
...
  },
  "components": {
    "schemas": {
...
      "ApproveDto": {
        "type": "object",
        "properties": {
          "rollbackNotes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
      "ArtifactDto": {
        "type": "object",
        "properties": {
          "kind": {
            "type": "string",
            "nullable": true
          },
          "sha256Hex": {
            "type": "string",
            "nullable": true
          },
          "sizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "storageRef": {
            "type": "string",
            "nullable": true
          },
          "metadataJson": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "CancelJobDto": {
        "type": "object",
        "properties": {
          "confirm": {
            "type": "boolean"
          },
          "reason": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false,
        "description": "Cancelling a job is audited — clients must send `confirm: true` (destructive / state-changing)."
      },
...
      "CompatibilityRuleDto": {
        "type": "object",
        "properties": {
          "requiredTerminalFirmwareVersionId": {
            "type": "string",
            "format": "uuid",
            "nullable": true
          },
          "requiredTargetSkuContains": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "FirmwareVersionCreateDto": {
        "type": "object",
        "properties": {
          "label": {
            "type": "string",
            "nullable": true
          },
          "buildId": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "PackageDto": {
        "type": "object",
        "properties": {
          "name": {
            "type": "string",
            "nullable": true
          },
          "versionLabel": {
            "type": "string",
            "nullable": true
          },
          "checksumHex": {
            "type": "string",
            "nullable": true
          },
          "artifactSizeBytes": {
            "type": "integer",
            "format": "int64"
          },
          "metadataJson": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
      "TargetDto": {
        "type": "object",
        "properties": {
          "firmwarePackageId": {
            "type": "string",
            "format": "uuid"
          },
          "sku": {
            "type": "string",
            "nullable": true
          },
          "notes": {
            "type": "string",
            "nullable": true
          }
        },
        "additionalProperties": false
      },
...
    }
  }
}